requirements.txt
Django==1.11.26
django-cors-headers==3.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
hashids==1.2.0
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
